move Payload/zBuf-related functionality into separate file
