+++ src/collections.rs
@@ -13,14 +13,7 @@
 //
 
 use libc::{c_char, size_t};
-use std::ops::Deref;
-use std::ptr::NonNull;
-use zenoh::{
-    buffers::{buffer::SplitBuffer, ZBuf},
-    prelude::ZenohId,
-};
-
-use crate::{impl_guarded_transmute, GuardedTransmute};
+use zenoh::prelude::ZenohId;
 
 /// A contiguous view of bytes owned by some other entity.
 ///
@@ -233,128 +226,5 @@ impl From<&[u8]> for z_bytes_t {
             start: s.as_ptr(),
             len: s.len(),
         }
-    }
-}
-
-pub use crate::z_owned_buffer_t;
-impl_guarded_transmute!(noderefs Option<ZBuf>, z_owned_buffer_t);
-impl Default for z_owned_buffer_t {
-    fn default() -> Self {
-        z_buffer_null()
-    }
-}
-impl From<ZBuf> for z_owned_buffer_t {
-    fn from(value: ZBuf) -> Self {
-        Some(value).transmute()
-    }
-}
-
-impl From<Option<ZBuf>> for z_owned_buffer_t {
-    fn from(value: Option<ZBuf>) -> Self {
-        match value {
-            Some(value) => value.into(),
-            None => z_buffer_null(),
-        }
-    }
-}
-impl core::ops::Deref for z_owned_buffer_t {
-    type Target = Option<ZBuf>;
-
-    fn deref(&self) -> &Self::Target {
-        unsafe { core::mem::transmute(self) }
-    }
-}
-impl core::ops::DerefMut for z_owned_buffer_t {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        unsafe { core::mem::transmute(self) }
-    }
-}
-
-/// The gravestone value for `z_owned_buffer_t`.
-#[no_mangle]
-pub extern "C" fn z_buffer_null() -> z_owned_buffer_t {
-    unsafe { core::mem::transmute(None::<ZBuf>) }
-}
-
-/// Decrements the buffer's reference counter, destroying it if applicable.
-///
-/// `buffer` will be reset to `z_buffer_null`, preventing UB on double-frees.
-#[no_mangle]
-pub extern "C" fn z_buffer_drop(buffer: &mut z_owned_buffer_t) {
-    core::mem::drop(buffer.take())
-}
-
-/// Returns `true` if the buffer is in a valid state.
-#[no_mangle]
-pub extern "C" fn z_buffer_check(buffer: &z_owned_buffer_t) -> bool {
-    buffer.is_some()
-}
-
-/// Loans the buffer, allowing you to call functions that only need a loan of it.
-#[no_mangle]
-pub extern "C" fn z_buffer_loan(buffer: &z_owned_buffer_t) -> z_buffer_t {
-    buffer.as_ref().into()
-}
-
-/// A loan of a `z_owned_buffer_t`.
-///
-/// As it is a split buffer, it may contain more than one slice. It's number of slices is returned by `z_buffer_slice_count`.
-#[repr(C)]
-#[derive(Clone, Copy, Default)]
-pub struct z_buffer_t {
-    _inner: Option<NonNull<z_owned_buffer_t>>,
-}
-
-impl From<Option<&ZBuf>> for z_buffer_t {
-    fn from(value: Option<&ZBuf>) -> Self {
-        unsafe { core::mem::transmute(value) }
-    }
-}
-
-impl From<z_buffer_t> for Option<&'static ZBuf> {
-    fn from(value: z_buffer_t) -> Self {
-        unsafe { core::mem::transmute(value) }
-    }
-}
-
-/// Increments the buffer's reference count, returning an owned version of the buffer.
-#[no_mangle]
-pub extern "C" fn z_buffer_clone(buffer: z_buffer_t) -> z_owned_buffer_t {
-    match buffer._inner {
-        Some(b) => unsafe { b.as_ref().deref().clone().transmute() },
-        None => ZBuf::empty().into(),
-    }
-}
-
-/// Returns the number of slices in the buffer.
-///
-/// If the return value is 0 or 1, then the buffer's data is contiguous in memory.
-#[no_mangle]
-pub extern "C" fn z_buffer_slice_count(buffer: z_buffer_t) -> usize {
-    match buffer.into() {
-        None => 0,
-        Some(buf) => ZBuf::slices(buf).len(),
-    }
-}
-
-/// Returns total number bytes in the buffer.
-#[no_mangle]
-pub extern "C" fn z_buffer_len(buffer: z_buffer_t) -> usize {
-    match buffer.into() {
-        None => 0,
-        Some(buf) => ZBuf::slices(buf).fold(0, |acc, s| acc + s.len()),
-    }
-}
-
-/// Returns the `index`th slice of the buffer, aliasing it.
-///
-/// Out of bounds accesses will return `z_bytes_empty`.
-#[no_mangle]
-pub extern "C" fn z_buffer_slice_at(buffer: z_buffer_t, index: usize) -> z_bytes_t {
-    match buffer.into() {
-        None => z_bytes_empty(),
-        Some(buf) => ZBuf::slices(buf)
-            .nth(index)
-            .map_or(z_bytes_empty(), |slice| slice.into()),
     }
 }
