+++ src/lib.rs
@@ -23,6 +23,8 @@ mod config;
 pub use crate::config::*;
 mod commons;
 pub use crate::commons::*;
+mod payload;
+pub use crate::payload::*;
 mod keyexpr;
 pub use crate::keyexpr::*;
 mod info;
